requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
pdfplumber
python-docx
nltk
matplotlib
wordcloud
